refactor(chat): extract web-search renderer + line primitive from ToolCallRenderer
The web-search expanded-result cluster (parse + action card + parsed-links
rendering) moves to webSearchRenderer.ts, and the shared renderLinesExpanded
primitive (used by 11 call sites) to toolLinesExpanded.ts so sibling renderers
reuse it without a cycle.

Behavior-preserving (ToolCallRenderer/SubagentRenderer/WriteEdit suites — 181
tests — pass untouched). ToolCallRenderer 766 -> 628 LOC; incidental
complexFunctions 238 -> 237 (the web-search dispatch left the file).
